Связь через http и tcp
Запросы на регистрацию и вход передаются по http
Игровая сессия проходит через tcp соединение
Клиент выдаёт статус входа в UI
Сервер не вылетает при отключении клиента
