+++ gameApp/util/data/userstatus/userstatus.go
@@ -1,0 +1,9 @@
+package userstatus
+
+type UserStatus uint8
+
+//go:generate stringer -type=UserStatus
+const (
+	Online UserStatus = iota + 1
+	Offline
+)
