--- servGame/compile.bat
@@ -1,2 +1,0 @@
-go build -o server.exe clientcon.go database.go gemeemu.go main.go server.go terminal.go
-server.exe
