+++ servGame/run.bat
@@ -1,0 +1,1 @@
+go run main.go
